# fern/definition/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Airweave",
    "version": "0.1.0",
    "description": "\n\n## API Groups\nThis API spec only includes the following API groups:\n- **Sources**: API endpoints for discovering available data source connectors and their configuration requirements\n- **Collections**: API endpoints for managing collections - logical groups of data sources that provide unified search capabilities\n- **Source Connections**: API endpoints for managing live connections to data sources. Source connections are the actual configured instances that Airweave uses to sync data from your apps and databases, transforming it into searchable, structured information within collections\n- **White Labels**: API endpoints for managing custom OAuth2 integrations with your own branding and credentials\n"
  },
  "paths": {
    "/sources/detail/{short_name}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Read Source",
        "description": "Get detailed information about a specific data source connector.",
        "operationId": "read_source_sources_detail__short_name__get",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "short_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Technical identifier of the source type (e.g., 'github', 'stripe', 'slack')",
              "title": "Short Name"
            },
            "description": "Technical identifier of the source type (e.g., 'github', 'stripe', 'slack')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Source"
                },
                "examples": {
                  "github_source": {
                    "summary": "Source details with authentication and configuration schemas",
                    "value": {
                      "id": "550e8400-e29b-41d4-a716-446655440000",
                      "name": "GitHub",
                      "description": "Connect to GitHub repositories for code, issues, pull requests, and documentation",
                      "auth_type": "config_class",
                      "auth_config_class": "GitHubAuthConfig",
                      "config_class": "GitHubConfig",
                      "short_name": "github",
                      "class_name": "GitHubSource",
                      "output_entity_definition_ids": [
                        "def12345-6789-abcd-ef01-234567890abc",
                        "def67890-abcd-ef01-2345-67890abcdef1"
                      ],
                      "labels": [
                        "code"
                      ],
                      "created_at": "2024-01-01T00:00:00Z",
                      "modified_at": "2024-01-01T00:00:00Z",
                      "auth_fields": {
                        "fields": [
                          {
                            "name": "personal_access_token",
                            "title": "Personal Access Token",
                            "description": "GitHub Personal Access Token with repository read permissions",
                            "type": "string",
                            "secret": true
                          },
                          {
                            "name": "repo_name",
                            "title": "Repository Name",
                            "description": "Full repository name in format 'owner/repo'",
                            "type": "string"
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sources/list": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Read Sources",
        "description": "List all available data source connectors.\n\n<br/><br/>\nReturns the complete catalog of source types that Airweave can connect to.",
        "operationId": "read_sources_sources_list_get",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Source"
                  },
                  "title": "Response Read Sources Sources List Get"
                },
                "examples": {
                  "available_sources": {
                    "summary": "List of all available data source connectors",
                    "value": [
                      {
                        "id": "550e8400-e29b-41d4-a716-446655440000",
                        "name": "GitHub",
                        "description": "Connect to GitHub repositories for code, issues, pull requests, and documentation",
                        "auth_type": "config_class",
                        "auth_config_class": "GitHubAuthConfig",
                        "config_class": "GitHubConfig",
                        "short_name": "github",
                        "class_name": "GitHubSource",
                        "output_entity_definition_ids": [
                          "def12345-6789-abcd-ef01-234567890abc",
                          "def67890-abcd-ef01-2345-67890abcdef1"
                        ],
                        "labels": [
                          "code"
                        ],
                        "created_at": "2024-01-01T00:00:00Z",
                        "modified_at": "2024-01-01T00:00:00Z",
                        "auth_fields": {
                          "fields": [
                            {
                              "name": "personal_access_token",
                              "title": "Personal Access Token",
                              "description": "GitHub Personal Access Token with repository read permissions",
                              "type": "string",
                              "secret": true
                            },
                            {
                              "name": "repo_name",
                              "title": "Repository Name",
                              "description": "Full repository name in format 'owner/repo'",
                              "type": "string"
                            }
                          ]
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthType": {
        "type": "string",
        "enum": [
          "oauth2",
          "oauth2_with_refresh",
          "oauth2_with_refresh_rotating",
          "api_key",
          "native_functionality",
          "config_class",
          "trello_auth",
          "none"
        ],
        "title": "AuthType",
        "description": "Enumeration of supported authentication types.\n\nAttributes:\n----------\n    oauth2: OAuth2 authentication.\n    oauth2_with_refresh: OAuth2 authentication with refresh token.\n    oauth2_with_refresh_rotating: OAuth2 authentication with rotating refresh token.\n    api_key: API key authentication.\n    native_functionality: Native functionality.\n    url_and_api_key: URL and API key authentication.\n    none: No authentication."
      },
...
      "ConfigField": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "required": {
            "type": "boolean",
            "title": "Required",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "name",
          "title",
          "type"
        ],
        "title": "ConfigField",
        "description": "Config field model."
      },
...
      "Fields": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/ConfigField"
            },
            "type": "array",
            "title": "Fields"
          }
        },
        "type": "object",
        "required": [
          "fields"
        ],
        "title": "Fields",
        "description": "Fields model."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Source": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Human-readable name of the data source connector (e.g., 'GitHub', 'Stripe', 'PostgreSQL')."
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Detailed description explaining what data this source can extract and its typical use cases."
          },
          "auth_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AuthType"
              },
              {
                "type": "null"
              }
            ],
            "description": "Type of authentication mechanism required by this source (e.g., 'oauth2')."
          },
          "auth_config_class": {
            "type": "string",
            "title": "Auth Config Class",
            "description": "Python class name that defines the authentication configuration fields required for this source."
          },
          "config_class": {
            "type": "string",
            "title": "Config Class",
            "description": "Python class name that defines the source-specific configuration options and parameters."
          },
          "short_name": {
            "type": "string",
            "title": "Short Name",
            "description": "Technical identifier used internally to reference this source type. Must be unique across all sources."
          },
          "class_name": {
            "type": "string",
            "title": "Class Name",
            "description": "Python class name of the source implementation that handles data extraction logic."
          },
          "output_entity_definition_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Entity Definition Ids",
            "description": "List of entity definition IDs that this source can produce. Defines the data schema and structure that this connector outputs."
          },
          "organization_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Organization Id",
            "description": "Organization identifier for custom source connectors. System sources have this set to null."
          },
          "labels": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Labels",
            "description": "Categorization tags to help users discover and filter sources by domain or use case."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique system identifier for this source type. Generated automatically when the source is registered."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Timestamp when this source type was registered in the system (ISO 8601 format)."
          },
          "modified_at": {
            "type": "string",
            "format": "date-time",
            "title": "Modified At",
            "description": "Timestamp when this source type was last updated (ISO 8601 format)."
          },
          "auth_fields": {
            "$ref": "#/components/schemas/Fields",
            "description": "Schema definition for authentication fields required to connect to this source. Describes field types, validation rules, and user interface hints."
          },
          "config_fields": {
            "$ref": "#/components/schemas/Fields",
            "description": "Schema definition for configuration fields required to customize this source. Describes field types, validation rules, and user interface hints."
          }
        },
        "type": "object",
        "required": [
          "name",
          "auth_config_class",
          "config_class",
          "short_name",
          "class_name",
          "id",
          "created_at",
          "modified_at",
          "auth_fields",
          "config_fields"
        ],
        "title": "Source",
        "description": "Complete source representation with authentication and configuration schemas.",
        "examples": [
          {
            "auth_config_class": "GitHubAuthConfig",
            "auth_fields": {
              "fields": [
                {
                  "description": "Personal Access Token with repository read permissions. Generate one at https://github.com/settings/tokens",
                  "name": "personal_access_token",
                  "secret": true,
                  "title": "Personal Access Token",
                  "type": "string"
                },
                {
                  "description": "Full repository name in format 'owner/repo' (e.g., 'airweave-ai/airweave')",
                  "name": "repo_name",
                  "title": "Repository Name",
                  "type": "string"
                }
              ]
            },
            "auth_type": "config_class",
            "class_name": "GitHubSource",
            "config_class": "GitHubConfig",
            "config_fields": {
              "fields": [
                {
                  "description": "Specific branch to sync (e.g., 'main', 'development'). If empty, uses the default branch.",
                  "name": "branch",
                  "title": "Branch name",
                  "type": "string"
                }
              ]
            },
            "created_at": "2024-01-01T00:00:00Z",
            "description": "Connect to GitHub repositories for code, issues, pull requests, and documentation",
            "id": "550e8400-e29b-41d4-a716-446655440000",
            "labels": [
              "code"
            ],
            "modified_at": "2024-01-01T00:00:00Z",
            "name": "GitHub",
            "output_entity_definition_ids": [
              "def12345-6789-abcd-ef01-234567890abc",
              "def67890-abcd-ef01-2345-67890abcdef1"
            ],
            "short_name": "github"
          }
        ]
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key for authentication"
      }
    }
  },
...
  "servers": [
    {
      "url": "https://api.airweave.ai",
      "description": "Production",
      "x-fern-server-name": "Production"
    },
    {
      "url": "http://localhost:8001",
      "description": "Local",
      "x-fern-server-name": "Local"
    }
  ]
}